video/FFv1: Decode slice/plane-rgb/lines
This commit implements the RLE-GR with bias cancellation and the plane decoding.

Information found in the headers (FrameHeader, slice headers) are used to find
the correct way to decode the slices.

If the colorspace type used is:
    * 0: decode the slice using the plane way
    * 1: decode the slice using the RGB way
    * other value: do not decode

Then, for each pixel of each line, decode it by using the context.

If the coder type is 0. use the adaptive decoding
If the coder type is 1, use the code range states with default value (127).
If the coder type is 2, use the code range states with custom values.

For the adaptive decoding, it uses the Run Length Encoding system using a symbol
encoded with the Golomb-Rice algorithm (modified with a bias correlation by context).

Some cleaning of the existing code is also performed.

Signed-off-by: Florent Tribouilloy <florent@mediaarea.net>