MKV/WebM: Report Delay and Duration correctly for VP9 MKV/WebM files.
No parser had been implemented for the VP9 codec, and the MKV analyzer was
not reporting Delay for any codec without a parser.

Since the parser is not actually necessary to extract the accurate Delay for
MKV video files, I did not fix this by implementing a VP9 parser; rather I
allowed the MKV analyzer to report a Delay if no parser is present.

With the Delay value detected correctly, the Duration can now be calculated
correctly as well.