Fix WM/EncodingTime parsing
The Windows Media parsing code tried to handle "WM/EncodingTime" but was failing because the value is a FILETIME, which uses units of 100 nanoseconds from 1601, but it was passing the value as-is to something that expects seconds from 1601. Just needs a division by 10000000 to fix.

Fix a side-effect of fixing, use seconds precision for both Header_FileProperties/CreationDate and WM/EncodingTime, compare them and show both only if they are different.